--- __mocks__/openmrs-esm-error-handling.mock.tsx
@@ -1,5 +1,0 @@
-import { never } from "rxjs";
-
-export function createErrorHandler() {
-  return jest.fn().mockReturnValue(never());
-}
